еще bug fixing
